Custom hook for better structure
