+++ assets/src/edit-story/components/library/libraryProvider.js
@@ -27,14 +27,8 @@ import { useState } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import {
-  DEFAULT_EDITOR_PAGE_WIDTH,
-  DEFAULT_EDITOR_PAGE_HEIGHT,
-} from '../../constants';
-import { createNewElement, getDefinitionForType } from '../../elements';
-import { editorToDataX, editorToDataY } from '../../units';
-import { useStory } from '../../app';
 import Context from './context';
+import useInsertElement from './useInsertElement';
 
 const MEDIA = 'media';
 const TEXT = 'text';
@@ -44,32 +38,7 @@ const ANIMATION = 'animation';
 
 function LibraryProvider({ children }) {
   const [tab, setTab] = useState(MEDIA);
-
-  const isMediaEl = (type) => {
-    const { isMedia } = getDefinitionForType(type);
-    return isMedia;
-  };
-  const {
-    actions: { addElement, setBackgroundElement },
-    state: { currentPage },
-  } = useStory();
-  const insertElement = (type, { width, height, ...props }) => {
-    const element = createNewElement(type, {
-      ...props,
-      x: editorToDataX(80 * Math.random(), DEFAULT_EDITOR_PAGE_WIDTH),
-      y: editorToDataY(70 * Math.random(), DEFAULT_EDITOR_PAGE_HEIGHT),
-      width: editorToDataX(width, DEFAULT_EDITOR_PAGE_WIDTH),
-      height: editorToDataY(height, DEFAULT_EDITOR_PAGE_HEIGHT),
-    });
-    addElement({ element });
-    if (
-      isMediaEl(type) &&
-      !currentPage.elements.some(({ type: elType }) => isMediaEl(elType))
-    ) {
-      setBackgroundElement({ elementId: element.id });
-    }
-    return element;
-  };
+  const insertElement = useInsertElement();
 
   const state = {
     state: {
